initial new layout changes
