+++ webapp/channels/src/components/admin_console/workspace-optimization/dashboard_checks/updates_data.ts
@@ -1,0 +1,3 @@
+// Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
+// See LICENSE.txt for license information.
+
