+++ webapp/channels/src/i18n/en.json
@@ -1949,14 +1949,13 @@
   "admin.reporting.workspace_optimization.configuration.descriptionOk": "You seem to have good configuration for SSL and Session Lengths!",
   "admin.reporting.workspace_optimization.configuration.session_length.description": "Your session length is set to the default of 30 days. A longer session length provides convenience, and a shorter session provides tighter security. We recommend adjusting this based on your organization's security policies.",
   "admin.reporting.workspace_optimization.configuration.session_length.title": "Session lengths is set to default",
+  "admin.reporting.workspace_optimization.configuration.session_length.cta": "Configure session length",
   "admin.reporting.workspace_optimization.configuration.ssl.description": "We strongly recommend securing your Mattermost workspace by configuring SSL in production environments.",
   "admin.reporting.workspace_optimization.configuration.ssl.title": "Configure SSL to make your server more secure",
-  "admin.reporting.workspace_optimization.configuration.title": "Configuration",
+  "admin.reporting.workspace_optimization.configuration.ssl.cta": "Configure web server", "admin.reporting.workspace_optimization.configuration.title": "Configuration",
   "admin.reporting.workspace_optimization.cta.configureDataRetention": "Try data retention",
   "admin.reporting.workspace_optimization.cta.configureElasticsearch": "Try Elasticsearch",
   "admin.reporting.workspace_optimization.cta.configureLDAP": "Try AD/LDAP",
-  "admin.reporting.workspace_optimization.cta.configureSessionLength": "Configure session length",
-  "admin.reporting.workspace_optimization.cta.configureWebServer": "Configure web server",
   "admin.reporting.workspace_optimization.cta.downloadUpdate": "Download update",
   "admin.reporting.workspace_optimization.cta.learnMore": "Learn more",
   "admin.reporting.workspace_optimization.cta.startTrial": "Start trial",
